+++ packages/skill-lesson/lib/sanity.ts
@@ -1,5 +1,6 @@
 import {
   CastingWordsOrderResponseSchema,
+  createCastingWordsOrder,
   getSRTText,
   getTranscriptText,
 } from './castingwords'
@@ -9,6 +10,9 @@ import groq from 'groq'
 import Mux from '@mux/mux-node'
 import {uniqueId} from 'lodash'
 import {sanityWriteClient} from '../utils/sanity-server'
+import {NextApiRequest, NextApiResponse} from 'next'
+import {createMuxAsset} from './mux'
+import {isValidSignature, SIGNATURE_HEADER_NAME} from '@sanity/webhook'
 
 const VideoResourceSchema = z.object({
   _id: z.string(),
@@ -137,4 +141,51 @@ export const updateVideoResourceWithTranscriptOrderId = async ({
       },
     })
     .commit()
+}
+
+const secret = process.env.SANITY_WEBHOOK_SECRET
+
+/**
+ * link to webhook {@link} https://www.sanity.io/organizations/om9qNpcXE/project/z9io1e0u/api/webhooks/xV5ZY6656qclI76i
+ *
+ * @param req
+ * @param res
+ */
+export const sanityVideoResourceWebhook = async (
+  req: NextApiRequest,
+  res: NextApiResponse,
+) => {
+  const signature = req.headers[SIGNATURE_HEADER_NAME] as string
+  const isValid = isValidSignature(JSON.stringify(req.body), signature, secret)
+
+  try {
+    if (isValid) {
+      const {_id, originalMediaUrl, castingwords, muxAsset, duration} = req.body
+      console.info('processing Sanity webhook: Video Resource created', _id)
+
+      const castingwordsOrder = await createCastingWordsOrder({
+        originalMediaUrl,
+        castingwords,
+      })
+
+      const {duration: assetDuration, ...newMuxAsset} = await createMuxAsset({
+        originalMediaUrl,
+        muxAsset,
+        duration,
+      })
+
+      await updateVideoResourceWithTranscriptOrderId({
+        sanityDocumentId: _id,
+        castingwordsOrder,
+        muxAsset: newMuxAsset,
+        duration: assetDuration,
+      })
+      res.status(200).json({success: true})
+    } else {
+      res.status(500).json({success: false})
+    }
+  } catch (e) {
+    Sentry.captureException(e)
+    res.status(200).json({success: true})
+  }
 }
