+++ packages/skill-lesson/package.json
@@ -34,6 +34,7 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-switch": "^1.0.1",
     "@sanity/client": "^6.1.1",
+    "@sanity/webhook": "^1.0.2",
     "@sentry/nextjs": "^7.9.0",
     "@skillrecordings/ability": "workspace:*",
     "@skillrecordings/analytics": "workspace:*",
